# bai-27-rest-api-1-2/api/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Book Store API
  version: 1.0.0
  contact:
    email: nguyenphucnguyenvy@gmail.com
servers:
  - url: https://1x20l.sse.codesandbox.io/api/v1
    description: BookStore API for dev
...
paths:
  /books:
    get:
      security:
        - BearerAuth: []
      summary: get books 
      tags:
        - books
      parameters: 
        - in: query
          name: page
          schema: 
            type: integer
            default: 0
            example: 0
        - in: query
          name: limit
          schema:
            type: integer
            default: 3
            example: 2
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  books: 
                    type: array
                    items: 
                      $ref: '#components/schemas/Book' 
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      length:
                        type: integer
                example:
                  books:
                    - _id: 5ea5700c9b99a306c9b0c268
                      title: Places Search Box
                      description: Fork from https://javascript.info/async-await
                      imgUrl: http://res.cloudinary.com/learn-nguyenvy/image/upload/v1587900422/student/805afbd0e48202972948e82b0c2c20d0.jpg
                      __v: 0
                    - _id: 5ea5701a9b99a306c9b0c269
                      title: Gauge series
                      description: Fork from https://javascript.info/async-await
                      imgUrl: http://res.cloudinary.com/learn-nguyenvy/image/upload/v1587900437/student/6612d02c1db62d1e6202d4b00c84f51f.jpg
                      __v: 0
                  pagination:
                    page: 0
                    limit: 2
                    length: 9
        '404':
          description: not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: Books not found
        '400':
          description: bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: Invalid request
    post:
      security:
        - BearerAuth: []   
      summary: add new book
      tags:
        - books
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title: 
                  type: string
                description: 
                  type: string
                avatar:
                  type: string
                  format: binary
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  newBook:
                    type: object
                    $ref: "#/components/schemas/Book"
        '400':
          description: not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: Invalid request
...
  /users:
    get:
      security:
        - BearerAuth: []
      summary: get users 
      tags:
        - users
      parameters: 
        - in: query
          name: page
          schema: 
            type: integer
            default: 0
            example: 0
        - in: query
          name: limit
          schema:
            type: integer
            default: 3
            example: 2
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  users: 
                    type: array
                    items: 
                      $ref: '#components/schemas/User' 
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      length:
                        type: integer
        '404':
          description: not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
    post:
      security:
        - BearerAuth: []     
      summary: add new user
      tags:
        - users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: 
                  type: string
                email: 
                  type: string
                  format: email
                imgUrl:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  newUser:
                    type: object
                    $ref: "#/components/schemas/User"
        '400':
          description: bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: Invalid request
...
  /transactions:
    get:
      security:
        - BearerAuth: []
      tags:
        - transactions
      summary: get transactions of store
      parameters:
        - in: query
          schema: 
            type: integer
            default: 0
          name: page
        - in: query
          schema: 
            type: integer
            default: 3
          name: limit
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: 
                type: object
                properties:
                  transactions:
                    type: object
                    properties:
                      _id: 
                        type: string
                      isCompleted:  
                        type: boolean
                      user:
                        $ref: '#/components/schemas/User'
                      books:
                        type: array
                        items:
                          $ref: '#/components/schemas/Book'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
    post:
      security:
        - BearerAuth: []
      tags:
        - transactions
      summary: add new transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                bookIds:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: 
                type: object
                properties:
                  transaction:
                    $ref: '#/components/schemas/Transaction'
        '400':
          description: bad request
          content:
            application/json:
              schema: 
                type: object
                properties:
                  message:
                    type: string             
...
components:
  schemas:
    Book:
      type: object
      properties:
        _id:
          type: string
        title:
          type: string
        description:
          type: string
        imgUrl:
          type: string
      example:
        _id: 5ea5700c9b99a306c9b0c268
        title: Places Search Box
        description: Fork from https://javascript.info/async-await
        imgUrl: http://res.cloudinary.com/learn-nguyenvy/image/upload/v1587900422/student/805afbd0e48202972948e82b0c2c20d0.jpg
        __v: 0
    User:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
        role:
          type: integer
          enum:
            - 0
            - 1
          default: 1
        isLogging:
          type: boolean
          default: false
        wrongLoginCount:
          type: integer
        avatar:
          type: string
      example:
        role: 0
        isLogging: true
        wrongLoginCount: 1
        _id: 5ea54ff9e3d8f9fe939a89b2
        name: nguyenvy
        password: $2b$10$X.wfHRypCk1Mb2si/jbEf.3/JipPtRH0utFr3YEqTRAtzcQ3ssqXq
        email: admin@gmail.com
        avatar: http://res.cloudinary.com/learn-nguyenvy/image/upload/v1587994173/student/1c287138f02a9fbe0abd01de94ab325b.png
    Transaction:
      type: object
      properties:
        userId:
          type: string
        bookIds:
          type: array
          items:
            type: string
        isCompleted:
          type: boolean
          default: false
      example:
        bookIds:
          - 5ea5700c9b99a306c9b0c268
          - 5ea5701a9b99a306c9b0c269
          - 5ea570299b99a306c9b0c26a
          - 5ea570419b99a306c9b0c26b
        isCompleted: false
        _id: 5eaaa57a76aed2005933b185
        userId: 5ea54ff9e3d8f9fe939a89b2
        __v: 0
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer